Rename `network.requester.request.headers.bytes` to 'memory'
