+++ core/mastodon/src/test/java/br/com/orcinus/orca/core/mastodon/network/requester/request/headers/memory/ByteReadChannelKSerializerTests.kt
@@ -13,31 +13,33 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package br.com.orcinus.orca.core.mastodon.network.requester.request.headers.bytes
+package br.com.orcinus.orca.core.mastodon.network.requester.request.headers.memory
 
 import assertk.assertThat
 import assertk.assertions.isEqualTo
-import java.nio.ByteBuffer
+import io.ktor.util.toByteArray
+import io.ktor.utils.io.ByteReadChannel
 import kotlin.test.Test
+import kotlinx.coroutines.test.runTest
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.add
 import kotlinx.serialization.json.buildJsonObject
 import kotlinx.serialization.json.putJsonArray
 
-internal class ByteBufferKSerializerTests {
+internal class ByteReadChannelKSerializerTests {
   @Test
   fun serializes() {
     assertThat(
         Json.encodeToString(
-          ByteBufferKSerializer,
-          ByteBuffer.wrap(byteArrayOf(0b000001, 0b00000010))
+          ByteReadChannel.serializer(),
+          ByteReadChannel(byteArrayOf(0b000001, 0b00000010))
         )
       )
       .isEqualTo(
         buildJsonObject {
             @OptIn(ExperimentalSerializationApi::class)
-            putJsonArray(ByteBufferKSerializer.descriptor.getElementName(0)) {
+            putJsonArray(ByteReadChannel.Companion.serializer().descriptor.getElementName(0)) {
               add(0b00000001)
               add(0b00000010)
             }
@@ -48,19 +50,24 @@ internal class ByteBufferKSerializerTests {
 
   @Test
   fun deserializes() {
-    assertThat(
-        Json.decodeFromString(
-          ByteBufferKSerializer,
-          buildJsonObject {
-              @OptIn(ExperimentalSerializationApi::class)
-              putJsonArray(ByteBufferKSerializer.descriptor.getElementName(0)) {
-                add(0b00000001)
-                add(0b00000010)
-              }
-            }
-            .toString()
+    runTest {
+      assertThat(
+          Json.decodeFromString(
+              ByteReadChannel.serializer(),
+              buildJsonObject {
+                  @OptIn(ExperimentalSerializationApi::class)
+                  putJsonArray(
+                    ByteReadChannel.Companion.serializer().descriptor.getElementName(0)
+                  ) {
+                    add(0b00000001)
+                    add(0b00000010)
+                  }
+                }
+                .toString()
+            )
+            .toByteArray()
         )
-      )
-      .isEqualTo(ByteBuffer.wrap(byteArrayOf(0b000001, 0b00000010)))
+        .isEqualTo(byteArrayOf(0b000001, 0b00000010))
+    }
   }
 }
